requirements.txt
numpy==1.23.3
opencv_python==4.6.0.66
PyYAML==6.0
